닉네임 변경 후 navbar 반영

+++ src/main/java/com/example/project1/member/service/MemberService.java
@@ -61,38 +61,54 @@ public class MemberService {
 
     }
 
-    public boolean remove(MemberForm data) {
-        Member member = memberRepository.findById(data.getId()).get();
+    public boolean remove(MemberForm data, MemberDto user) {
+        if (user != null) {
+            Member member = memberRepository.findById(data.getId()).get();
+            if (member.getId().equals(user.getId())) {
+                String dbPw = member.getPassword();
+                String formPw = data.getPassword();
 
-        String dbPw = member.getPassword();
-        String formPw = data.getPassword();
-        if (dbPw.equals(formPw)) {
-            memberRepository.delete(member);
-            return true;
-        } else {
-            return false;
+                if (dbPw.equals(formPw)) {
+                    memberRepository.delete(member);
+                    return true;
+                }
+            }
         }
+        return false;
     }
 
-    public boolean update(MemberForm data) {
-        // 조회
-        Member member = memberRepository.findById(data.getId()).get();
+    public boolean update(MemberForm data, MemberDto user, HttpSession session) {
+        if (user != null) {
+            // 조회
+            Member member = memberRepository.findById(data.getId()).get();
+            if (member.getId().equals(user.getId())) {
+                String dbPw = member.getPassword();
+                String formPw = data.getPassword();
 
-        String dbPw = member.getPassword();
-        String formPw = data.getPassword();
+                if (dbPw.equals(formPw)) {
+                    // 변경
+                    member.setNickname(data.getNickName());
+                    member.setInfo(data.getInfo());
+                    // 저장
+                    memberRepository.save(member);
 
-        if (dbPw.equals(formPw)) {
-            // 변경
-            member.setNickname(data.getNickName());
-            member.setInfo(data.getInfo());
-            // 저장
-            memberRepository.save(member);
-
-            return true;
-        } else {
-            return false;
+                    // dto 만들어서 session 에 저장
+                    addUserToSession(session, member);
+                    return true;
+                }
+            }
         }
+        return false;
+    }
 
+    private static void addUserToSession(HttpSession session, Member member) {
+        MemberDto dto = new MemberDto();
+        dto.setId(member.getId());
+        dto.setNickName(member.getNickname());
+        dto.setInfo(member.getInfo());
+        dto.setCreatedAt(member.getCreatedAt());
+
+        session.setAttribute("loggedInUser", dto);
     }
 
     public boolean updatePassword(String id, String oldPassword, String newPassword) {
@@ -118,13 +134,7 @@ public class MemberService {
             if (dbPassword.equals(password)) {
 
                 // memberDto 를 session에 넣기
-                MemberDto dto = new MemberDto();
-                dto.setId(db.get().getId());
-                dto.setNickName(db.get().getNickname());
-                dto.setInfo(db.get().getInfo());
-                dto.setCreatedAt(db.get().getCreatedAt());
-
-                session.setAttribute("loggedInUser", dto);
+                addUserToSession(session, db.get());
                 return true;
             }
         }
